test(domain): 폴더 UseCase 테스트 FolderUseCaseTest로 통합
- CreateFolderUseCaseTest, FetchFolderUseCaseTest, UpdateFolderUseCaseTest를 FolderUseCaseTest 하나로 통합
- Presentation 테스트(FolderViewModelTests, MainViewModelTests, OnBoardingViewModelTests)에서 FolderUseCase 직접 사용으로 변경
