+++ Presentation/Tests/Folder/FolderViewModelTests.swift
@@ -41,8 +41,7 @@ final class FolderViewModelTests: XCTestCase {
 
         let viewModel = FolderViewModel(
             category: initialCategory,
-            createUseCase: DefaultCreateFolderUseCase(repository: mockFolderRepo),
-            updateUseCase: DefaultUpdateFolderUseCase(repository: mockFolderRepo),
+            folderUseCase: DefaultFolderUseCase(repository: mockFolderRepo),
             wasteBasketRepository: mockWasteBasketRepo
         )
         viewModel.coordinator = mockCoordinator
